Moved /index.json to new ci.webapi blueprint
